% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster_utils.R
\name{build_contiguous_raster}
\alias{build_contiguous_raster}
\title{Build contiguous network raster (patches + corridors)}
\usage{
build_contiguous_raster(habitat_mask, corridor_raster, connectivity = 8)
}
\arguments{
\item{habitat_mask}{SpatRaster of habitat.}

\item{corridor_raster}{SpatRaster of corridors.}

\item{connectivity}{Connectivity for patches (4 or 8).}
}
\value{
SpatRaster with component sizes.
}
\description{
Build contiguous network raster (patches + corridors)
}
