% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster_utils.R
\name{build_patch_candidates}
\alias{build_patch_candidates}
\title{Build candidate edges between patches (centroid distance)}
\usage{
build_patch_candidates(patch_df, max_search_distance, raster_ref)
}
\arguments{
\item{patch_df}{Patch summary data frame.}

\item{max_search_distance}{Maximum distance in pixels.}

\item{raster_ref}{Raster reference for pixel size.}
}
\value{
Data frame with patch1, patch2, cost, distance_map, id.
}
\description{
Build candidate edges between patches (centroid distance)
}
