% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_metrics.R
\name{calculate_disturbance_penalty}
\alias{calculate_disturbance_penalty}
\title{Calculate disturbance penalty (diameter normalized)}
\usage{
calculate_disturbance_penalty(graph)
}
\arguments{
\item{graph}{igraph graph.}
}
\value{
Numeric penalty.
}
\description{
Calculate disturbance penalty (diameter normalized)
}
