% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_metrics.R
\name{calculate_failure_probability}
\alias{calculate_failure_probability}
\title{Estimate failure probability via edge removal}
\usage{
calculate_failure_probability(graph, k_failures = 1, iterations = 100)
}
\arguments{
\item{graph}{igraph graph.}

\item{k_failures}{Number of edges removed each trial.}

\item{iterations}{Number of Monte Carlo iterations.}
}
\value{
Failure probability.
}
\description{
Estimate failure probability via edge removal
}
