% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_metrics.R
\name{calculate_movement_entropy}
\alias{calculate_movement_entropy}
\title{Calculate movement entropy for a graph}
\usage{
calculate_movement_entropy(graph, alpha = 0.002)
}
\arguments{
\item{graph}{igraph graph.}

\item{alpha}{Dispersal kernel parameter.}
}
\value{
Numeric entropy value.
}
\description{
Calculate movement entropy for a graph
}
