% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_metrics.R
\name{calculate_topology_penalty}
\alias{calculate_topology_penalty}
\title{Calculate topology penalty (cycle count)}
\usage{
calculate_topology_penalty(graph)
}
\arguments{
\item{graph}{igraph graph.}
}
\value{
Numeric penalty.
}
\description{
Calculate topology penalty (cycle count)
}
