% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_metrics.R
\name{calculate_total_entropy}
\alias{calculate_total_entropy}
\title{Calculate total entropy summary}
\usage{
calculate_total_entropy(graph, lambda_c = 1, lambda_f = 1, lambda_d = 1)
}
\arguments{
\item{graph}{igraph graph.}

\item{lambda_c}{Connectivity penalty multiplier.}

\item{lambda_f}{Topology penalty multiplier.}

\item{lambda_d}{Disturbance penalty multiplier.}
}
\value{
Named list of entropy components.
}
\description{
Calculate total entropy summary
}
