% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_metrics.R
\name{calculate_two_edge_connectivity}
\alias{calculate_two_edge_connectivity}
\title{Calculate the fraction of node pairs with two edge-disjoint paths}
\usage{
calculate_two_edge_connectivity(graph)
}
\arguments{
\item{graph}{igraph graph.}
}
\value{
Numeric ratio.
}
\description{
Calculate the fraction of node pairs with two edge-disjoint paths
}
