% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster_utils.R
\name{label_patches}
\alias{label_patches}
\title{Label contiguous habitat patches}
\usage{
label_patches(mask, connectivity = 8)
}
\arguments{
\item{mask}{Logical SpatRaster mask.}

\item{connectivity}{4 or 8.}
}
\value{
SpatRaster of patch labels.
}
\description{
Label contiguous habitat patches
}
