% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster_utils.R
\name{raster_mask_from_values}
\alias{raster_mask_from_values}
\title{Create a logical mask from raster values}
\usage{
raster_mask_from_values(raster, values)
}
\arguments{
\item{raster}{SpatRaster.}

\item{values}{Numeric values to keep.}
}
\value{
Logical SpatRaster mask.
}
\description{
Create a logical mask from raster values
}
