% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terralink-package.R
\docType{package}
\name{terralink-package}
\alias{terralink}
\alias{terralink-package}
\title{terralink: Connectivity Corridor Optimization}
\description{
Standalone R implementation of habitat connectivity corridor optimization
algorithms for raster and vector workflows, without desktop GIS
dependencies.
}
\details{
Main user entry points:
\itemize{
\item \code{terralink_raster()} for raster inputs
\item \code{terralink_vector()} for polygon patch inputs
\item \code{terralink_run()} for a single generic wrapper
}

Raster inputs are funneled through TerraLink's vector corridor pipeline so
the R package matches the current QGIS plugin workflow.
}
\seealso{
\itemize{
\item \code{\link{terralink_raster}}
\item \code{\link{terralink_vector}}
\item \code{\link{terralink_run}}
\item \url{https://github.com/sorus-tools/terralink-r}
\item Report bugs at \url{https://github.com/sorus-tools/terralink-r/issues}
}
}
\author{
\strong{Maintainer}: Benjamin Bishop \email{benjamin.bishop@sorusconsultingllc.com}

Other contributors:
\itemize{
  \item SORUS Consulting LLC [funder, copyright holder]
}

}
