% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engine.R
\name{UnionFind}
\alias{UnionFind}
\title{Union-Find data structure}
\description{
Union-Find data structure

Union-Find data structure
}
\section{Methods}{

\describe{
\item{initialize}{Create a new UnionFind.}
\item{find}{Find root of a node with path compression.}
\item{union}{Union two nodes; returns TRUE if merged.}
\item{get_size}{Get component size for a node.}
\item{get_count}{Get component count for a node.}
}
}

\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{parent}}{Environment mapping nodes to parents.}

\item{\code{size}}{Environment mapping roots to component sizes.}

\item{\code{count}}{Environment mapping roots to component counts.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-UnionFind-new}{\code{UnionFind$new()}}
\item \href{#method-UnionFind-find}{\code{UnionFind$find()}}
\item \href{#method-UnionFind-union}{\code{UnionFind$union()}}
\item \href{#method-UnionFind-get_size}{\code{UnionFind$get_size()}}
\item \href{#method-UnionFind-get_count}{\code{UnionFind$get_count()}}
\item \href{#method-UnionFind-clone}{\code{UnionFind$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UnionFind-new"></a>}}
\if{latex}{\out{\hypertarget{method-UnionFind-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UnionFind$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UnionFind-find"></a>}}
\if{latex}{\out{\hypertarget{method-UnionFind-find}{}}}
\subsection{Method \code{find()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UnionFind$find(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{Node id for lookup operations.}

\item{\code{x}}{Node id for lookup operations.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UnionFind-union"></a>}}
\if{latex}{\out{\hypertarget{method-UnionFind-union}{}}}
\subsection{Method \code{union()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UnionFind$union(a, b)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{a}}{Node id for union operations.}

\item{\code{a}}{Node id for union operations.}

\item{\code{b}}{Node id for union operations.}

\item{\code{b}}{Node id for union operations.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UnionFind-get_size"></a>}}
\if{latex}{\out{\hypertarget{method-UnionFind-get_size}{}}}
\subsection{Method \code{get_size()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UnionFind$get_size(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{Node id for lookup operations.}

\item{\code{x}}{Node id for lookup operations.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UnionFind-get_count"></a>}}
\if{latex}{\out{\hypertarget{method-UnionFind-get_count}{}}}
\subsection{Method \code{get_count()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UnionFind$get_count(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{Node id for lookup operations.}

\item{\code{x}}{Node id for lookup operations.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UnionFind-clone"></a>}}
\if{latex}{\out{\hypertarget{method-UnionFind-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UnionFind$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
