% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terralink-package.R
\docType{package}
\name{terralink-package}
\alias{terralink}
\alias{terralink-package}
\title{terralink: Connectivity Corridor Optimization}
\description{
Standalone R implementation of habitat connectivity corridor optimization
algorithms for raster and vector workflows, without desktop GIS
dependencies.
}
\seealso{
\itemize{
\item \url{https://github.com/sorus-tools/terralink-r}
\item Report bugs at \url{https://github.com/sorus-tools/terralink-r/issues}
}
}
\author{
\strong{Maintainer}: Benjamin Bishop \email{info@sorusconsultingllc.com}

Other contributors:
\itemize{
  \item SORUS Consulting LLC [funder, copyright holder]
}

}
\keyword{internal}
