% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terralink-package.R
\docType{package}
\name{terralink-package}
\alias{terralink}
\alias{terralink-package}
\title{terralink: TerraLink Corridor Optimization}
\description{
Standalone R implementation of the TerraLink habitat connectivity corridor optimization algorithms, ported from the QGIS plugin without QGIS dependencies.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/sorus-tools/TerraLink}
  \item Report bugs at \url{https://github.com/sorus-tools/TerraLink/issues}
}

}
\author{
\strong{Maintainer}: Benjamin Bishop \email{benjamin.bishop@sorusconsultingllc.com}

}
\keyword{internal}
